cleanup + ipgeo fix
